+++ src/main/java/kitchenpos/eatinorders/domain/JpaOrderTableRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.eatinorders.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+}
