+++ src/main/java/kitchenpos/eatinorders/domain/NumberOfGuests.java
@@ -1,0 +1,33 @@
+package kitchenpos.eatinorders.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class NumberOfGuests {
+    @Column(name = "number_of_guests", nullable = false)
+    private int value;
+
+    protected NumberOfGuests() {
+    }
+
+    public NumberOfGuests(int value) {
+        validate(value);
+        this.value = value;
+    }
+
+    public void changeNumberOfGuests(int value) {
+        validate(value);
+        this.value = value;
+    }
+
+    private void validate(int value) {
+        if (value < 0) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public int getValue() {
+        return value;
+    }
+}
