+++ src/main/java/kitchenpos/eatinorders/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorders.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
