+++ src/main/java/kitchenpos/eatinorders/domain/Quantity.java
@@ -1,0 +1,21 @@
+package kitchenpos.eatinorders.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class Quantity {
+    @Column(name = "quantity", nullable = false)
+    private long value;
+
+    protected Quantity() {
+    }
+
+    public Quantity(long quantity) {
+        this.value = quantity;
+    }
+
+    public long getValue() {
+        return value;
+    }
+}
