--- src/main/java/kitchenpos/eatinorders/tobe/application/OrderValidator.java
@@ -1,48 +1,0 @@
-package kitchenpos.eatinorders.tobe.application;
-
-import kitchenpos.eatinorders.tobe.application.dto.request.OrderLineItemCreateRequest;
-import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.domain.MenuRepository;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Objects;
-import java.util.stream.Collectors;
-
-@Component
-public class OrderValidator {
-    private final MenuRepository menuRepository;
-
-    public OrderValidator(final MenuRepository menuRepository) {
-        this.menuRepository = menuRepository;
-    }
-
-    public void validateMenu(OrderLineItemCreateRequest orderLineItemRequest) {
-        final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
-            .orElseThrow(NoSuchElementException::new);
-        if (!menu.isDisplayed()) {
-            throw new IllegalStateException();
-        }
-        if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public static void validateOrderLineItemRequests(List<OrderLineItemCreateRequest> orderLineItemRequests) {
-        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void validateExistMenu(List<OrderLineItemCreateRequest> orderLineItemRequests) {
-        final List<Menu> menus = menuRepository.findAllByIdIn(
-            orderLineItemRequests.stream()
-                .map(OrderLineItemCreateRequest::getMenuId)
-                .collect(Collectors.toList())
-        );
-        if (menus.size() != orderLineItemRequests.size()) {
-            throw new IllegalArgumentException();
-        }
-    }
-}
