--- src/main/java/kitchenpos/eatinorders/tobe/application/dto/request/OrderTableChangeRequest.java
@@ -1,16 +1,0 @@
-package kitchenpos.eatinorders.tobe.application.dto.request;
-
-public class OrderTableChangeRequest {
-    private int numberOfGuests;
-
-    public OrderTableChangeRequest() {
-    }
-
-    public OrderTableChangeRequest(int numberOfGuests) {
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
