--- src/main/java/kitchenpos/eatinorders/tobe/domain/OrderLineItems.java
@@ -1,33 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-import javax.persistence.CascadeType;
-import javax.persistence.Embeddable;
-import javax.persistence.ForeignKey;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-@Embeddable
-public class OrderLineItems {
-    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
-    @JoinColumn(
-        name = "order_id",
-        nullable = false,
-        columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_order_line_item_to_orders")
-    )
-    private List<OrderLineItem> orderLineItems = new ArrayList<>();
-
-    protected OrderLineItems() {
-    }
-
-    public void addAll(List<OrderLineItem> orderLineItems) {
-        this.orderLineItems.addAll(orderLineItems);
-    }
-
-    public List<OrderLineItem> get() {
-        return Collections.unmodifiableList(orderLineItems);
-    }
-}
