--- src/main/java/kitchenpos/eatinorders/tobe/domain/OrderTableName.java
@@ -1,29 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import java.util.Objects;
-
-@Embeddable
-public class OrderTableName {
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    protected OrderTableName() {
-    }
-
-    public OrderTableName(String name) {
-        validate(name);
-        this.name = name;
-    }
-
-    private void validate(String name) {
-        if (Objects.isNull(name) || name.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public String getValue() {
-        return name;
-    }
-}
