--- src/main/java/kitchenpos/eatinorders/tobe/domain/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
