--- src/main/java/kitchenpos/eatinorders/tobe/domain/Price.java
@@ -1,34 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import java.math.BigDecimal;
-import java.util.Objects;
-
-@Embeddable
-public class Price {
-    @Column(name = "price", nullable = false)
-    private BigDecimal value;
-
-    protected Price() {
-    }
-
-    public Price(BigDecimal value) {
-        validate(value);
-        this.value = value;
-    }
-
-    private void validate(BigDecimal price) {
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public BigDecimal getValue() {
-        return value;
-    }
-
-    public kitchenpos.products.domain.Price multiply(BigDecimal price) {
-        return new kitchenpos.products.domain.Price(this.getValue().multiply(price));
-    }
-}
