--- src/main/java/kitchenpos/eatinorders/tobe/domain/Quantity.java
@@ -1,21 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Quantity {
-    @Column(name = "quantity", nullable = false)
-    private long value;
-
-    protected Quantity() {
-    }
-
-    public Quantity(long quantity) {
-        this.value = quantity;
-    }
-
-    public long getValue() {
-        return value;
-    }
-}
