--- src/main/java/kitchenpos/eatinorders/tobe/ui/OrderTableRestController.java
@@ -1,58 +1,0 @@
-package kitchenpos.eatinorders.tobe.ui;
-
-import kitchenpos.eatinorders.tobe.application.OrderTableService;
-import kitchenpos.eatinorders.tobe.application.dto.request.OrderTableChangeRequest;
-import kitchenpos.eatinorders.tobe.application.dto.request.OrderTableCreateRequest;
-import kitchenpos.eatinorders.tobe.application.dto.response.OrderTableResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-import java.util.UUID;
-
-@RequestMapping("/api/order-tables")
-@RestController
-public class OrderTableRestController {
-    private final OrderTableService orderTableService;
-
-    public OrderTableRestController(final OrderTableService orderTableService) {
-        this.orderTableService = orderTableService;
-    }
-
-    @PostMapping
-    public ResponseEntity<OrderTableResponse> create(@RequestBody final OrderTableCreateRequest request) {
-        final OrderTableResponse response = orderTableService.create(request);
-        return ResponseEntity.created(URI.create("/api/order-tables/" + response.getId()))
-            .body(response);
-    }
-
-    @PutMapping("/{orderTableId}/sit")
-    public ResponseEntity<OrderTableResponse> sit(@PathVariable final UUID orderTableId) {
-        return ResponseEntity.ok(orderTableService.sit(orderTableId));
-    }
-
-    @PutMapping("/{orderTableId}/clear")
-    public ResponseEntity<OrderTableResponse> clear(@PathVariable final UUID orderTableId) {
-        return ResponseEntity.ok(orderTableService.clear(orderTableId));
-    }
-
-    @PutMapping("/{orderTableId}/number-of-guests")
-    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
-        @PathVariable final UUID orderTableId,
-        @RequestBody final OrderTableChangeRequest request
-    ) {
-        return ResponseEntity.ok(orderTableService.changeNumberOfGuests(orderTableId, request));
-    }
-
-    @GetMapping
-    public ResponseEntity<List<OrderTableResponse>> findAll() {
-        return ResponseEntity.ok(orderTableService.findAll());
-    }
-}
